Admin middleware

+++ examen/routes/web.php
@@ -24,57 +24,40 @@ Route::get('/over-DTV', function () {
 Auth::routes();
 
 //home user page
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home')->middleware('auth');
 
 //admin---------------------------------
 //admin index
-Route::get('/admin', function () {
-    return view('Backend.AdminIndex');
-});
-Route::get('/adminlogin', function () {
-    return view('Backend.AdminLogin');
-});
-
-Route::get('/pref',
-    function () {
-
-        return view('Backend.preferences');
-    });
-
+Route::get('/admin', function () {return view('Backend.AdminIndex');})->middleware('admin')->name('adminIndex');
+Route::get('/adminlogin', function () {return view('Backend.AdminLogin');})->middleware('admin');
+Route::get('/pref',function () {return view('Backend.preferences');})->middleware('admin');
 
 //banen---------------------------------
 //baan index
-Route::get('/adminbaan', [App\Http\Controllers\BaanController::class, 'index'])->name('adminbaans');
+Route::get('/adminbaan', [App\Http\Controllers\BaanController::class, 'index'])->name('adminbaans')->middleware('admin');
 //baan create
-Route::get('/adminbaan-create', [App\Http\Controllers\BaanController::class, 'indexCreate'])->name('createbaan');
-Route::post('/adminbaan-create', [App\Http\Controllers\BaanController::class, 'create'])->name('createpostbaan');
+Route::get('/adminbaan-create', [App\Http\Controllers\BaanController::class, 'indexCreate'])->name('createbaan')->middleware('admin');
+Route::post('/adminbaan-create', [App\Http\Controllers\BaanController::class, 'create'])->name('createpostbaan')->middleware('admin');
 //baan edit $id
-Route::get('/adminbaan-edit/{id}', [App\Http\Controllers\BaanController::class, 'edit'])->name('editbaan');
-Route::post('/adminbaan-edit/{id}', [App\Http\Controllers\BaanController::class, 'update'])->name('editpostbaan');
+Route::get('/adminbaan-edit/{id}', [App\Http\Controllers\BaanController::class, 'edit'])->name('editbaan')->middleware('admin');
+Route::post('/adminbaan-edit/{id}', [App\Http\Controllers\BaanController::class, 'update'])->name('editpostbaan')->middleware('admin');
 //baan delete
-Route::delete('/adminbaan{baan}', [App\Http\Controllers\BaanController::class, 'destroy'])->name('destroybaan');
-
+Route::delete('/adminbaan{baan}', [App\Http\Controllers\BaanController::class, 'destroy'])->name('destroybaan')->middleware('admin');
 
 //users---------------------------------
 //user index
-Route::get('/usermanage', [App\Http\Controllers\UserController::class, 'index'])->name('userindex');
+Route::get('/usermanage', [App\Http\Controllers\UserController::class, 'index'])->name('userindex')->middleware('admin');
 //user delete
-Route::delete('/usermanage{User}', [App\Http\Controllers\UserController::class, 'destroy'])->name('userdelete');
+Route::delete('/usermanage{User}', [App\Http\Controllers\UserController::class, 'destroy'])->name('userdelete')->middleware('admin');
 //edit
-Route::get('/usermanage{users}', [App\Http\Controllers\UserController::class, 'edit'])->name('useredit');
-Route::put('/usermanage{users}', [App\Http\Controllers\UserController::class, 'update'])->name('userpostedit');
-
+Route::get('/usermanage{users}', [App\Http\Controllers\UserController::class, 'edit'])->name('useredit')->middleware('admin');
+Route::put('/usermanage{users}', [App\Http\Controllers\UserController::class, 'update'])->name('userpostedit')->middleware('admin');
 
 //toernooien---------------------------
-Route::get('/toernooien', [App\Http\Controllers\ToernooiController::class, 'index'])->name('toernooiindex');
+Route::get('/toernooien', [App\Http\Controllers\ToernooiController::class, 'index'])->name('toernooiindex')->middleware('admin');
 //toernooi create
-Route::get('/toernooi-create', [App\Http\Controllers\ToernooiController::class, 'store'])->name('createtoernooi');
-Route::post('/toernooi-create', [App\Http\Controllers\ToernooiController::class, 'create'])->name('createposttoernooi');
-
-
-
-
-
+Route::get('/toernooi-create', [App\Http\Controllers\ToernooiController::class, 'store'])->name('createtoernooi')->middleware('admin');
+Route::post('/toernooi-create', [App\Http\Controllers\ToernooiController::class, 'create'])->name('createposttoernooi')->middleware('admin');
 
 //Frontend Kantine
 //Show
